Helper.go function added
