+++ components/automate-cli/cmd/chef-automate/infrastructure.go
@@ -2,10 +2,7 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"os"
-	"os/exec"
-	"time"
 
 	api "github.com/chef/automate/api/interservice/deployment"
 	"github.com/chef/automate/components/automate-cli/pkg/docs"
@@ -16,6 +13,18 @@ import (
 	"github.com/spf13/cobra"
 	"google.golang.org/grpc"
 )
+
+var nodeDeleteCmd = &cobra.Command{
+	Use:               "node-delete [uuid]",
+	Short:             "Delete node by node uuid",
+	Long:              "",
+	PersistentPreRunE: checkLicenseStatusForExpiry,
+	RunE:              runDeleteNodeCmd,
+	Args:              cobra.ExactArgs(1),
+	Annotations: map[string]string{
+		docs.Tag: docs.BastionHost,
+	},
+}
 
 func init() {
 	infrastructureCmd.AddCommand(nodeDeleteCmd)
@@ -33,18 +42,6 @@ var infrastructureCmd = &cobra.Command{
 	},
 }
 
-var nodeDeleteCmd = &cobra.Command{
-	Use:               "node-delete [uuid]",
-	Short:             "Delete node by node uuid",
-	Long:              "",
-	PersistentPreRunE: checkLicenseStatusForExpiry,
-	RunE:              runDeleteNodeCmd,
-	Args:              cobra.ExactArgs(1),
-	Annotations: map[string]string{
-		docs.Tag: docs.BastionHost,
-	},
-}
-
 type DSClient interface {
 	InfrastructureNodeDelete(ctx context.Context, in *api.InfrastructureNodeDeleteRequest, opts ...grpc.CallOption) (*api.InfrastructureNodeDeleteResponse, error)
 	Close() error
@@ -53,24 +50,6 @@ type DSClient interface {
 type InfraFlow struct {
 	DsClient DSClient
 	Writer   *cli.Writer
-}
-
-type LicenseResult struct {
-	Result           LicenseStatus `json:"result"`
-	ErrorType        string        `json:"error_type"`
-	ErrorDescription string        `json:"error_description"`
-}
-
-type LicenseStatus struct {
-	CustomerName   string         `json:"customer_name"`
-	LicenseType    string         `json:"license_type"`
-	LicenseId      string         `json:"license_id"`
-	ExpirationDate ExpirationDate `json:"expiration_date"`
-	GracePeriod    bool           `json:"grace_period"`
-}
-
-type ExpirationDate struct {
-	Seconds int64 `json:"seconds"`
 }
 
 func runDeleteNodeCmd(cmd *cobra.Command, args []string) error {
@@ -95,90 +74,6 @@ func preInfrastructureCmd(cmd *cobra.Command, args []string) error {
 		os.Exit(0)
 	}
 	return nil
-}
-
-func checkLicenseStatusForExpiry(cmd *cobra.Command, args []string) error {
-	fileName := "/tmp/license"
-	err := commandPrePersistent(cmd)
-	if err != nil {
-		return status.Wrap(err, status.CommandExecutionError, "unable to set command parent settings")
-	}
-
-	cmd1 := exec.Command("chef-automate", "license", "status", "--result-json", fileName)
-	//Not checking for error, as if the license is expired it will still return error and for commercial license a grace period is required
-	cmd1.Output()
-
-	licenseResult, err := readFileAndMarshal(fileName)
-	if err != nil {
-		return err
-	}
-
-	err = checkLicenseExpiry(licenseResult)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func readFileAndMarshal(fileName string) (*LicenseResult, error) {
-	var result LicenseResult
-
-	byteValue, err := os.ReadFile(fileName)
-	if err != nil {
-		return nil, err
-	}
-
-	defer os.Remove(fileName)
-	json.Unmarshal([]byte(byteValue), &result)
-
-	return &result, nil
-
-}
-
-func checkLicenseExpiry(licenseResult *LicenseResult) error {
-	commercial := "commercial"
-	if licenseResult.Result.LicenseId == "" {
-		if licenseResult.ErrorType != "" {
-			return status.New(
-				status.DeploymentServiceCallError,
-				licenseResult.ErrorDescription,
-			)
-		}
-		return status.New(
-			status.LicenseError,
-			"Please apply a license.Please contact sales@chef.io to have your Chef Automate license.",
-		)
-	}
-	licenseValidDate := time.Unix(licenseResult.Result.ExpirationDate.Seconds, 0) //gives unix time stamp in utc
-
-	//If the license type is commercial, adding grace period of 1 month
-	licenseResult.Result.GracePeriod = false
-	if licenseResult.Result.LicenseType == commercial {
-		// Check if the license is expired
-		if licenseValidDate.Before(time.Now()) {
-			// Calculate the end date of the grace period (30 days after the license expiry)
-			gracePeriodEnd := licenseValidDate.AddDate(0, 0, 30)
-
-			//checks if current date and time is within the grace period
-			if gracePeriodEnd.After(time.Now()) {
-				//If within the grace period, set GracePeriod to true and extend license validity
-				licenseResult.Result.GracePeriod = true
-				licenseValidDate = gracePeriodEnd
-			}
-		}
-	}
-
-	// Check if the license (including the grace period) is expired
-	if licenseValidDate.Before(time.Now()) {
-		return status.New(
-			status.LicenseError,
-			"This license has expired and grace period is also got over. Please contact sales@chef.io to renew your Chef Automate license.",
-		)
-	}
-
-	return nil
-
 }
 
 func NewDeleteNode(writer *cli.Writer) (*InfraFlow, error) {
